school/common/address.py
from hospital.apps.web.home.models import Address


def get_address(self):

    address = Address.objects.\
        values('code','phone_number', 'email1', 'street_name',\
            'house_number', 'post_code', 'area', 'city', 'region', 'country' )\
                .first()

    return  address