requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==10.0.0
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0